calibrate mag for 10 seconds only
we don't want to run this all the time, as our algo isn't particularly
robust; it's just a least-squares fit, so if you do something funky like
the IMU next to a big magnet, the squared error goes nuts and the
calibration goes out the window.

so for now, we're just going to wave it around for 10 seconds, and then
be done with it.